backend/pyproject.toml
[tool.poetry]
name = "restful api crud"
version = "0.1.0"
description = "Task App Exam"
authors = ["Ezequiel Bravo"]


[tool.poetry.dependencies]
python = "^3.12"
annotated-types = "0.7.0"
anyio = "4.4.0"
build = "1.2.1"
bump-pydantic = "0.8.0"
cachecontrol = "0.14.0"
certifi = "2024.6.2"
charset-normalizer = "3.3.2"
cleo = "2.1.0"
click = "8.1.7"
colorama = "0.4.6"
crashtest = "0.4.1"
distlib = "0.3.8"
dnspython = "2.6.1"
dulwich = "0.21.7"
email-validator = "2.2.0"
fastapi = "0.111.0"
fastapi-cli = "0.0.4"
fastjsonschema = "2.20.0"
filelock = "3.15.4"
flake8-plugin-utils = "1.3.3"
flake8-pytest-style = "2.0.0"
h11 = "0.14.0"
httpcore = "1.0.5"
httptools = "0.6.1"
httpx = "0.27.0"
idna = "3.7"
iniconfig = "2.0.0"
installer = "0.7.0"
jaraco-classes = "3.4.0"
jinja2 = "3.1.4"
keyring = "24.3.1"
libcst = "1.4.0"
markdown-it-py = "3.0.0"
markupsafe = "2.1.5"
mdurl = "0.1.2"
mongomock = "4.1.2"
more-itertools = "10.3.0"
motor = "3.5.0"
msgpack = "1.0.8"
orjson = "3.10.5"
packaging = "24.1"
pexpect = "4.9.0"
pkginfo = "1.11.1"
platformdirs = "4.2.2"
pluggy = "1.5.0"
poetry = "1.8.3"
poetry-core = "1.9.0"
poetry-plugin-export = "1.8.0"
ptyprocess = "0.7.0"
pydantic = "2.7.4"
pydantic-core = "2.18.4"
pygments = "2.18.0"
pymongo = "4.8.0"
pyproject-hooks = "1.1.0"
pytest = "8.2.2"
python-decouple = "3.8"
python-dotenv = "1.0.1"
python-multipart = "0.0.9"
pywin32-ctypes = "0.2.2"
pyyaml = "6.0.1"
rapidfuzz = "3.9.3"
requests = "2.32.3"
requests-toolbelt = "1.0.0"
rich = "13.7.1"
sentinels = "1.0.0"
setuptools = "70.1.1"
shellingham = "1.5.4"
sniffio = "1.3.1"
starlette = "0.37.2"
tomlkit = "0.12.5"
trove-classifiers = "2024.5.22"
typer = "0.12.3"
typing-extensions = "4.12.2"
ujson = "5.10.0"
urllib3 = "2.2.2"
uvicorn = "0.30.1"
virtualenv = "20.26.3"
watchfiles = "0.22.0"
websockets = "12.0"
flake8 = "^7.1.0"
pytest-asyncio = "^0.23.7"
asyncio = "^3.4.3"
aiohttp = "^3.9.5"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
